c2-il: name the SECOND blocker of every receiver form, and count how many
D2 attached a whole-body-complete count to each of 29 receiver forms and ranked
two rungs off it correctly, but three forms holding 172,625 functions read 0.0 %
and a column that is zero for its three largest rows has no ordering information
left. §14.6 said so explicitly and left the question open: what else blocks those
bodies.

The completeness matcher now walks past the receiver form, names the construct
that blocks the body next, and grants constructs greedily — up to four — so the
key carries the pair a rung would actually have to implement plus how many
constructs the body needs in all. Still decode without acceptance: every path
returns a refusal.

Six new constructs, each with a witness: a conditional branch (its label is
defined by a later `29 <same token>`), a chain link, a bare CALL token over an
already-pushed callee, a byte-offset add outside a receiver designator, a
pointer-typed operand in an argument region, and the by-value temporary bind.
`walk` also tokenizes a `99` bind met inside an open argument region, without
which a member call nested in an argument list files as `op-0x99`.

Attribution is by construct, never by position: the refusal keeps the *furthest*
point reached, a `26` there is re-classified by `walk`, and a call whose receiver
is the wrong designator carries the `26`'s own offset — otherwise a decayed string
literal, a named object and a by-value temporary become three uninformative
opcodes. `Block::aux` widens to u64 rather than being squeezed, because
truncating an intrinsic selector would merge two census buckets. The opcode-name
table moves to `body::expr_opcode_name` so the two key families cannot disagree
about what a byte is called.

`chained` gains the production D2 left unwritten, and the residue tail is 24
distinct opcode bytes over 4.2 % of the bucket with 80 % of that in one known
byte — not the flat 197-bucket shape that means a payload is being read as
vocabulary.

Acceptance unchanged: census 280,020/2,462,571, per-TU in-class and per-TU class
identical on all 878 TUs, every non-mcall census key and sum identical, four mode
lanes 45/42/42/42 with 0 mismatch, bench 110/0/0, expr_sweep checked=3259
mismatches=0. The sub-buckets sum to D2's parents exactly; 19 functions are
re-decomposed out of `op-0x99` and accounted for one by one.

+++ crates/c2-il/src/func/body/expr.rs
@@ -306,7 +306,7 @@ pub(crate) fn parse_expr(seg: &[u8], p: &mut usize, stop: u8) -> Result<Vec<IlOp
                 ctx: "expr-intrinsic",
                 byte: Some(0x40),
                 off: *p,
-                aux: id as u32,
+                aux: id as u64,
             });
         }
         match b {
